feat: Nuke Trajectory predicts with alliance breakage
